requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
groq>=0.32.0
gspread>=6.2.1
google-auth>=2.41.1
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0
python-dotenv>=1.1.1
streamlit-authenticator>=0.3.2
